// app/2026/data/types.ts
export enum RegionalStatus {
  preparing = "preparing",
  preliminariesFinished = "preliminariesFinished",
  regionalsFinished = "regionalsFinished",
}

export interface RegionalBase {
  contestDate: `2025-${string}-${string}`;
  site: string;
  region: string;
  url?: string;
  score: number;
  scoreOverride?: number;
  scoreDetails: RegionScoreArgs;
  disclaimer?: ReactNode;
  lastYear?: Regional2025;
}

export interface TeamInStandingsLike {
  teamId: number;
  rank: number | null;
  problemsSolved: number;
  institution: string;
  teamName: string;
}
// ...
export interface TeamInTeamListLikeTeamIdNullable {
  teamId: number | null;
  name: string;
  institution: string;
  problemsSolved?: number;
}

export type Regional = RegionalBase &
  (
    | {
        status: RegionalStatus.preliminariesFinished;
        regionalTeams: TeamInTeamListLikeTeamIdNullable[];
      }
    | {
        status: RegionalStatus.regionalsFinished;
        scoreboard: TeamInStandingsLike[];
      }
    | { status: RegionalStatus.preparing }
  );